+++ docs/changelog.md
@@ -4,4 +4,4 @@
 
 ## v0.1.0 (06/12/2023)
 
-- First release of `aiod_sdk`!
+- First release of `aiondemand`!